Add test for Label Only and restructure inference test module
