+++ tests/attacks/inference/attribute_inference/test_black_box.py
@@ -25,14 +25,9 @@ import torch.nn as nn
 import torch.optim as optim
 
 from art.attacks.inference.attribute_inference.black_box import AttributeInferenceBlackBox
-from art.attacks.inference.attribute_inference.white_box_decision_tree import AttributeInferenceWhiteBoxDecisionTree
-from art.attacks.inference.attribute_inference.white_box_lifestyle_decision_tree import (
-    AttributeInferenceWhiteBoxLifestyleDecisionTree,
-)
 from art.estimators.classification.pytorch import PyTorchClassifier
 from art.estimators.estimator import BaseEstimator
 from art.estimators.classification import ClassifierMixin
-from art.estimators.classification.scikitlearn import ScikitlearnDecisionTreeClassifier
 
 from tests.attacks.utils import backend_test_classifier_type_check_fail
 from tests.utils import ARTTestException
@@ -148,67 +143,5 @@ def test_black_box_with_model(art_warning, decision_tree_estimator, get_iris_dat
         art_warning(e)
 
 
-@pytest.mark.skipMlFramework("dl_frameworks")
-def test_white_box(art_warning, decision_tree_estimator, get_iris_dataset):
-    try:
-        attack_feature = 2  # petal length
-        values = [0.14, 0.42, 0.71]  # rounded down
-        priors = [50 / 150, 54 / 150, 46 / 150]
-
-        (x_train_iris, y_train_iris), (x_test_iris, y_test_iris) = get_iris_dataset
-        x_train_for_attack = np.delete(x_train_iris, attack_feature, 1)
-        x_train_feature = x_train_iris[:, attack_feature]
-        x_test_for_attack = np.delete(x_test_iris, attack_feature, 1)
-        x_test_feature = x_test_iris[:, attack_feature]
-
-        classifier = decision_tree_estimator()
-
-        attack = AttributeInferenceWhiteBoxDecisionTree(classifier, attack_feature=attack_feature)
-        x_train_predictions = np.array([np.argmax(arr) for arr in classifier.predict(x_train_iris)]).reshape(-1, 1)
-        x_test_predictions = np.array([np.argmax(arr) for arr in classifier.predict(x_test_iris)]).reshape(-1, 1)
-        inferred_train = attack.infer(x_train_for_attack, x_train_predictions, values=values, priors=priors)
-        inferred_test = attack.infer(x_test_for_attack, x_test_predictions, values=values, priors=priors)
-        train_diff = np.abs(inferred_train - x_train_feature.reshape(1, -1))
-        test_diff = np.abs(inferred_test - x_test_feature.reshape(1, -1))
-        assert np.sum(train_diff) / len(inferred_train) == pytest.approx(0.2108, abs=0.03)
-        assert np.sum(test_diff) / len(inferred_test) == pytest.approx(0.1988, abs=0.03)
-    except ARTTestException as e:
-        art_warning(e)
-
-
-@pytest.mark.skipMlFramework("dl_frameworks")
-def test_white_box_lifestyle(art_warning, decision_tree_estimator, get_iris_dataset):
-    try:
-        attack_feature = 2  # petal length
-        values = [0.14, 0.42, 0.71]  # rounded down
-        priors = [50 / 150, 54 / 150, 46 / 150]
-
-        (x_train_iris, y_train_iris), (x_test_iris, y_test_iris) = get_iris_dataset
-        x_train_for_attack = np.delete(x_train_iris, attack_feature, 1)
-        x_train_feature = x_train_iris[:, attack_feature]
-        x_test_for_attack = np.delete(x_test_iris, attack_feature, 1)
-        x_test_feature = x_test_iris[:, attack_feature]
-
-        classifier = decision_tree_estimator()
-        attack = AttributeInferenceWhiteBoxLifestyleDecisionTree(classifier, attack_feature=attack_feature)
-        x_train_predictions = np.array([np.argmax(arr) for arr in classifier.predict(x_train_iris)]).reshape(-1, 1)
-        x_test_predictions = np.array([np.argmax(arr) for arr in classifier.predict(x_test_iris)]).reshape(-1, 1)
-        inferred_train = attack.infer(x_train_for_attack, x_train_predictions, values=values, priors=priors)
-        inferred_test = attack.infer(x_test_for_attack, x_test_predictions, values=values, priors=priors)
-        train_diff = np.abs(inferred_train - x_train_feature.reshape(1, -1))
-        test_diff = np.abs(inferred_test - x_test_feature.reshape(1, -1))
-        assert np.sum(train_diff) / len(inferred_train) == pytest.approx(0.3357, abs=0.03)
-        assert np.sum(test_diff) / len(inferred_test) == pytest.approx(0.3149, abs=0.03)
-        # assert np.sum(train_diff) / len(inferred_train) < np.sum(test_diff) / len(inferred_test)
-    except ARTTestException as e:
-        art_warning(e)
-
-
 def test_classifier_type_check_fail():
     backend_test_classifier_type_check_fail(AttributeInferenceBlackBox, (BaseEstimator, ClassifierMixin))
-    backend_test_classifier_type_check_fail(
-        AttributeInferenceWhiteBoxLifestyleDecisionTree, (ScikitlearnDecisionTreeClassifier,)
-    )
-    backend_test_classifier_type_check_fail(
-        AttributeInferenceWhiteBoxDecisionTree, (ScikitlearnDecisionTreeClassifier,)
-    )
